# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_key}/board/move": {
      "post": {
        "tags": [
          "routes::board"
        ],
        "operationId": "move_issue",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_key}/sprints": {
      "get": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "list_sprints",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "create_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_key}/sprints/{sprint_id}": {
      "get": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "get_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "update_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_key}/sprints/{sprint_id}/close": {
      "post": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "close_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_key}/sprints/{sprint_id}/issues": {
      "post": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "move_issue_to_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveIssueToSprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_key}/sprints/{sprint_id}/remove-issue": {
      "post": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "remove_issue_from_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveIssueToSprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_key}/sprints/{sprint_id}/start": {
      "post": {
        "tags": [
          "routes::sprints"
        ],
        "operationId": "start_sprint",
        "parameters": [
          {
            "name": "project_key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprint_id",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardColumnResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "issue_ids"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "issue_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "wip_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "BoardResponse": {
        "type": "object",
        "required": [
          "project_id",
          "project_key",
          "columns",
          "issues",
          "sprint"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnResponse"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            }
          },
          "project_id": {
            "type": "string"
          },
          "project_key": {
            "type": "string"
          },
          "sprint": {
            "$ref": "#/components/schemas/SprintResponse"
          }
        }
      },
...
      "CreateSprintRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "goal": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "IssueResponse": {
        "type": "object",
        "required": [
          "id",
          "key",
          "summary",
          "description",
          "issue_type",
          "project_key",
          "status",
          "status_id",
          "priority",
          "labels",
          "reporter_id",
          "project_name"
        ],
        "properties": {
          "assignee_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "assignee_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "issue_type": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "original_estimate_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Original estimate in seconds. Null when no estimate was set."
          },
          "priority": {
            "type": "string"
          },
          "project_key": {
            "type": "string"
          },
          "project_name": {
            "type": "string"
          },
          "remaining_estimate_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Remaining estimate in seconds, derived from the latest worklog."
          },
          "reporter_id": {
            "type": "string"
          },
          "reporter_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "sprint_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "status_id": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "MoveIssueRequest": {
        "type": "object",
        "required": [
          "issue_id",
          "status_id"
        ],
        "properties": {
          "issue_id": {
            "type": "string"
          },
          "status_id": {
            "type": "string"
          }
        }
      },
      "MoveIssueToSprintRequest": {
        "type": "object",
        "required": [
          "issue_id"
        ],
        "properties": {
          "issue_id": {
            "type": "string"
          }
        }
      },
...
      "SprintListResponse": {
        "type": "object",
        "required": [
          "sprints"
        ],
        "properties": {
          "sprints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SprintResponse"
            }
          }
        }
      },
      "SprintResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "goal",
          "state",
          "velocity",
          "issue_ids"
        ],
        "properties": {
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Format: date-time"
          },
          "goal": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "issue_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "remaining_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Format: date-time"
          },
          "state": {
            "type": "string"
          },
          "velocity": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateSprintRequest": {
        "type": "object",
        "properties": {
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "goal": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  }
}